backend/pyproject.toml
[project]
name = "ai-ads-manager"
version = "0.1.0"
description = "AI-powered ad management for Wildberries and Ozon"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "asyncpg>=0.30.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "celery[redis]>=5.4.0",
    "httpx>=0.28.0",
    "redis>=5.0.0",
    "openai>=1.0.0",
    "alembic>=1.14.0",
    "cryptography>=44.0.0",
    "python-dotenv>=1.0.0",
    "aiofiles>=24.0.0",
    "pandas>=2.2.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.25.0",
    "pytest-httpx>=0.35.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py313"
